moved modules from "xcube.im" into "xcube.core" and "xcube.util"
