+++ test/util/test_ndarray_to_cmap_image.py
