+++ test/util/test_tilegrid.py
@@ -1,7 +1,7 @@
 from unittest import TestCase
 
-from xcube.webapi.im.tilegrid import GLOBAL_GEO_EXTENT
-from xcube.webapi.im.tilegrid import TileGrid, pow2_1d_subdivisions, pow2_2d_subdivision
+from xcube.util.tilegrid import GLOBAL_GEO_EXTENT
+from xcube.util.tilegrid import TileGrid, pow2_1d_subdivisions, pow2_2d_subdivision
 
 
 class TilingSchemeTest(TestCase):
