+++ xcube/core/timeseries.py
@@ -26,9 +26,9 @@ import shapely.geometry
 import shapely.wkt
 import xarray as xr
 
-from ..core.select import select_vars
-from ..core.verify import assert_cube
 from xcube.core.geom import mask_dataset_by_geometry, convert_geometry, GeometryLike, get_dataset_geometry
+from xcube.core.select import select_vars
+from xcube.core.verify import assert_cube
 
 Date = Union[np.datetime64, str]
